server/requirements.txt
django>=4.0
psycopg>=2.9.7
gunicorn>=20.1.0
django-configurations
whitenoise==6.5.0
djangorestframework
markdown
django-filter
drf-yasg
django-cities
Pillow
djangorestframework-gis